Clean up config a bit
