+++ components/config/src/config/languages.rs
@@ -1,0 +1,22 @@
+use std::collections::HashMap;
+
+use serde_derive::{Deserialize, Serialize};
+
+#[derive(Clone, Debug, PartialEq, Eq, Serialize, Deserialize)]
+#[serde(default)]
+pub struct Language {
+    /// The language code
+    pub code: String,
+    /// Whether to generate a feed for that language, defaults to `false`
+    pub feed: bool,
+    /// Whether to generate search index for that language, defaults to `false`
+    pub search: bool,
+}
+
+impl Default for Language {
+    fn default() -> Self {
+        Language { code: String::new(), feed: false, search: false }
+    }
+}
+
+pub type TranslateTerm = HashMap<String, String>;
